requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
supabase==2.3.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
gunicorn==21.2.0